% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{all_bromeliads_list_to_df}
\alias{all_bromeliads_list_to_df}
\title{Converte bromeliad list to a dataframe}
\usage{
all_bromeliads_list_to_df(.all_bromeliads_list)
}
\arguments{
\item{.all_bromeliads_list}{list where each element is all the bromeliads of
a certain visit.}
}
\value{
a dataframe, one row per bromeliad
}
\description{
also does a little checking to make sure that the visit IDs of the bromeliads
that come back are UNIQUE and identical with what was asked for.
}
\details{
Also tells  you how many bromeliads you've got.
}
