% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{get_all_abundances}
\alias{get_all_abundances}
\title{Get all Abundances from a list of datasets}
\usage{
get_all_abundances(.dats)
}
\arguments{
\item{.dats}{a dataframe containing datasets. usually the result of using
`bwg_get("datasets")`}
}
\value{
a dataframe with one row per invertebrate species found in each
  dataset. abundances across all bromeliads are in a NESTED dataframe to
  reduce clutter.
}
\description{
Get all Abundances from a list of datasets
}
