% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{get_all_bromeliads}
\alias{get_all_bromeliads}
\title{Get all bromeliad information from the database}
\usage{
get_all_bromeliads(.dats)
}
\arguments{
\item{.dats}{a dataframe containing visits. it must have a column called
"visit_id" which stores the IDs for the visits whose bromeliads will be
downloaded. Typically these would be unique, as in the output from
`bwg_get("visits")`.}
}
\value{
a list with one element per row of the input dataframe. Each element
  it named after a visit_id, and contains a dataframe of all the bromeliads
  collected during that visit
}
\description{
Get all bromeliad information from the database
}
