% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-abundance.R
\name{make_brom_a_df}
\alias{make_brom_a_df}
\title{Process abundance data for bromeliads and convert to a data frame}
\usage{
make_brom_a_df(abd_data)
}
\arguments{
\item{abd_data}{A nested list containing abundance data.}
}
\value{
A nested data frame where each element corresponds to the 
  abundance of a species in a particular bromeliad.
}
\description{
This function processes abundance data at a depth of 2 in a nested list 
and returns a data frame for each dataset. It extracts measurements related 
to bromeliads and calculates abundance values.
}
