% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-abundance.R
\name{one_measurement}
\alias{one_measurement}
\title{Convert a list of measurements into a single data frame}
\usage{
one_measurement(measure_list)
}
\arguments{
\item{measure_list}{A list containing measurements for a particular insect 
size.}
}
\value{
A data frame representing the measurements for a specific insect size.
}
\description{
This function processes a list of measurements for a specific insect size 
and returns them as a flat data frame.
}
