% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-abundance.R
\name{test_numeric}
\alias{test_numeric}
\title{Test if numeric values are valid}
\usage{
test_numeric(to_num)
}
\arguments{
\item{to_num}{A vector of values to be converted to numeric.}
}
\value{
A numeric vector where invalid values are converted to NA. 
  It also updates the 'weird' environment with the invalid entries.
}
\description{
This function converts input values into numeric format and stores any 
values that are not valid numbers (NA) into a list of problems.
}
