% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-abundance.R
\name{tidy_dataset_list}
\alias{tidy_dataset_list}
\title{Tidy a list of abundance data in to a data.frame}
\usage{
tidy_dataset_list(matrix_data_list)
}
\arguments{
\item{matrix_data_list}{}
}
\description{
This is a function that runs many smaller functions. together, they process 
data from the output of \code{bwg_get("matrix")}. The function assumes that
there is a list of such datasets. It relies heavily on \code{purrr}
}

