% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{get_all_bromeliads}
\alias{get_all_bromeliads}
\title{Get all bromeliad information from the database}
\usage{
get_all_bromeliads(.dats)
}
\arguments{
\item{.dats}{a dataframe containing visits. it must have a column called
"visit_id" which stores the IDs for the visits whose bromeliads will be
downloaded. Typically these would be unique, as in the output from
`bwg_get("visits")`.}
}
\value{
a single dataframe with a column called visit_id, where visit ID is
  stored as an integer. The dataframe has one row per bromeliad.
}
\description{
Get all bromeliad information from the database
}
